test(wasm): the typed write and resolve suites run on the gated surface
`runtime/runtime.js` patches `writer`/`reader` onto the core build's `Quill`.
`basic.test.js` drives the typst backend build, a different class over
different memory, where those verbs do not exist — which is why its
typed-commit and resolve suites reached for `_commitField` and `_resolve`.
Both suites move to `runtime.test.js`, where the gated surface is what a
consumer holds, and no `_`-prefixed call remains in either file.

Five assertions that `runtime.test.js` already made through `writer.set` and
`writer.card(i).set` are gone. Ten had no public twin and carry over:
`edit::field_coercion_failed` and `edit::field_not_inline`, which appeared
nowhere else in the suite; the `DocPath` a refused write anchors to;
`setAll`'s all-or-nothing abort at both the unknown-name and the coercion
rung, now one test; `CardWriter.setAll`; and resolve's declaration-order
rows, its default and blank rungs, and `body` as a sibling of `fields`.

Six tests filed under the typed-commit title never touched that ABI — they
drive `applyChange` and `mapMarks` — and are retitled where they sit.

The foreign-module fixture copies the core build into an `mkdtemp` directory
and removes it in `afterAll`. A unique path per run cannot collide with a
concurrent run, and a leak lands in the OS temp directory rather than in
`pkg/`.
