+++ crates/bindings/wasm/basic.test.js
@@ -812,82 +812,13 @@ Kindless card.
   })
 })
 
-// The typed-commit ABI is `_commitField` / `_commitFields` (hidden from the
-// `.d.ts`); `quill.writer(doc)` delegates here. These exercise the ABI directly.
-describe('Document typed-commit ABI: _commitField / _commitFields', () => {
-  const COMMIT_QUILL_YAML = `quill:
-  name: commit_test
-  version: "1.0"
-  backend: typst
-  description: Typed write smoke test
-
-main:
-  fields:
-    subject:
-      type: richtext
-      inline: true
-    intro:
-      type: richtext
-    qty:
-      type: integer
-
-card_kinds:
-  note:
-    fields:
-      body:
-        type: richtext
-`
-  const buildQuill = () =>
-    Quill.fromTree(makeQuill({ name: 'commit_test', plate: TEST_PLATE, quillYaml: COMMIT_QUILL_YAML }))
+// One `applyChange` bundle carries three channels — a text delta, island ops
+// and line ops — over a field or body whose anchors survive the splice, and
+// `mapMarks` answers where each mark lands before the bundle is applied.
+describe('Document applyChange: the anchor-preserving change bundle', () => {
   const blankDoc = () => Document.fromMarkdown('~~~card-yaml\n$quill: commit_test\n~~~\n\nBody.')
 
-  it('commitField resolves the schema type: richtext string → content, integer "3" → 3', () => {
-    const quill = buildQuill()
-    const doc = blankDoc()
-    doc._commitField(quill, 'intro', 'A **bold** intro.')
-    expect(typeof field(doc.main, 'intro')).toBe('object')
-    // The markdown projection of a richtext field is exportMarkdown ∘ its content.
-    expect(exportMarkdown(field(doc.main, 'intro'))).toBe('A **bold** intro.')
-
-    doc._commitField(quill, 'qty', '3')
-    expect(field(doc.main, 'qty')).toBe(3)
-  })
-
-  it('commitField rejects an unknown field as a typo and writes nothing', () => {
-    const quill = buildQuill()
-    const doc = blankDoc()
-    expectEditCode(() => doc._commitField(quill, 'stray', 'x'), 'edit::unknown_field')
-    expect(hasField(doc.main, 'stray')).toBe(false)
-    // Opaque storage stays available on purpose through the raw verb.
-    doc.storeField('stray', 'x')
-    expect(field(doc.main, 'stray')).toBe('x')
-  })
-
-  it('exportMarkdown composes on a committed richtext field; a scalar field is not a content', () => {
-    const quill = buildQuill()
-    const doc = blankDoc()
-    // Absent field: the value is undefined, nothing to project.
-    expect(field(doc.main, 'nonexistent')).toBeUndefined()
-    // A non-richtext scalar is stored verbatim, not a content object.
-    doc.storeField('count', 3)
-    expect(field(doc.main, 'count')).toBe(3)
-    // A committed richtext field projects through the codec.
-    doc._commitField(quill, 'intro', 'plain intro')
-    expect(exportMarkdown(field(doc.main, 'intro'))).toBe('plain intro')
-  })
-
-  it('commitField fails a strict mismatch and a richtext(inline) violation', () => {
-    const quill = buildQuill()
-    const doc = blankDoc()
-    expectEditCode(() => doc._commitField(quill, 'qty', 'not-a-number'), 'edit::field_coercion_failed')
-    expectEditCode(
-      () => doc._commitField(quill, 'subject', 'line one\n\nline two'),
-      'edit::field_not_inline',
-    )
-  })
-
   it('revise({field}) rebases a richtext field anchor and applyChange splices it', () => {
-    const quill = buildQuill()
     const doc = blankDoc()
     // revise the field from markdown (edit semantics), then splice a formatting
     // mark over "bold" via applyChange.
@@ -1045,87 +976,6 @@ card_kinds:
     expect(doc.main.body.islands.length).toBe(1)
   })
 
-  it('commitCardField resolves the card-kind schema and errors on a bad index', () => {
-    const quill = buildQuill()
-    const doc = Document.fromMarkdown(
-      '~~~card-yaml\n$quill: commit_test\n~~~\n\nMain.\n\n~~~card-yaml\n$kind: note\n~~~\n\nCard.',
-    )
-    doc._commitField(quill, { card: 0, field: 'body' }, 'Card **body**.')
-    expect(exportMarkdown(field(doc.cards[0], 'body'))).toBe('Card **body**.')
-    expectEditCode(() => doc._commitField(quill, { card: 9, field: 'body' }, 'x'), 'edit::index_out_of_range')
-  })
-
-  it('a mutator diagnostic carries the DocPath it anchors to', () => {
-    const quill = buildQuill()
-    const doc = Document.fromMarkdown(
-      '~~~card-yaml\n$quill: commit_test\n~~~\n\nMain.\n\n~~~card-yaml\n$kind: note\n~~~\n\nCard.',
-    )
-    // The path a diagnostic thrown by `fn` carries.
-    const pathOf = (fn) => {
-      try {
-        fn()
-      } catch (err) {
-        return err.diagnostics[0].path
-      }
-      throw new Error('expected a throw, got none')
-    }
-    // A main field conform error anchors at the rooted main field DocPath…
-    expect(pathOf(() => doc._commitField(quill, 'qty', 'not-a-number'))).toBe('main.qty')
-    // …a card field is kind-qualified with its absolute index…
-    expect(pathOf(() => doc._commitField(quill, { card: 0, field: 'stray' }, 'x'))).toBe(
-      'cards.note[0].stray',
-    )
-    // …and a structural out-of-range op anchors at the array slot.
-    expect(pathOf(() => doc.moveCard(9, 0))).toBe('cards[9]')
-    // `pathFor` mints what the anchor carries, so a consumer's path and the
-    // engine's agree without a kind table of its own.
-    expect(doc.pathFor({ card: 0, field: 'stray' })).toBe(
-      pathOf(() => doc._commitField(quill, { card: 0, field: 'stray' }, 'x')),
-    )
-  })
-
-  it('commitFields typed-commits a batch', () => {
-    const quill = buildQuill()
-    const doc = blankDoc()
-    doc._commitFields(quill, {}, { intro: 'A **bold** intro.', qty: '3' })
-    // The values were coerced, not stored verbatim.
-    expect(exportMarkdown(field(doc.main, 'intro'))).toBe('A **bold** intro.')
-    expect(field(doc.main, 'qty')).toBe(3)
-  })
-
-  it('commitFields aborts the whole batch on a typo, reporting the unknown field', () => {
-    const quill = buildQuill()
-    const doc = blankDoc()
-    // `qty` is a schema field; `titel` is a typo the schema does not own: the
-    // undeclared name aborts the all-or-nothing batch and nothing is applied.
-    expectEditCode(() => doc._commitFields(quill, {}, { qty: '5', titel: 'oops' }), 'edit::unknown_field')
-    expect(hasField(doc.main, 'qty')).toBe(false)
-    expect(hasField(doc.main, 'titel')).toBe(false)
-  })
-
-  it('commitFields is all-or-nothing: a bad field aborts the whole batch', () => {
-    const quill = buildQuill()
-    const doc = blankDoc()
-    // `subject` is richtext(inline); a multi-block value violates it, so nothing
-    // is applied: `qty` must not linger.
-    expectEditCode(
-      () => doc._commitFields(quill, {}, { qty: '5', subject: 'line one\n\nline two' }),
-      'edit::field_not_inline',
-    )
-    expect(hasField(doc.main, 'qty')).toBe(false)
-  })
-
-  it('commitCardFields typed-commits card fields and errors on a bad index', () => {
-    const quill = buildQuill()
-    const doc = Document.fromMarkdown(
-      '~~~card-yaml\n$quill: commit_test\n~~~\n\nMain.\n\n~~~card-yaml\n$kind: note\n~~~\n\nCard.',
-    )
-    doc._commitFields(quill, { card: 0 }, { body: 'Card **body**.' })
-    expect(exportMarkdown(field(doc.cards[0], 'body'))).toBe('Card **body**.')
-    // An undeclared field on the card aborts the batch.
-    expectEditCode(() => doc._commitFields(quill, { card: 0 }, { stray: 'x' }), 'edit::unknown_field')
-    expectEditCode(() => doc._commitFields(quill, { card: 9 }, { body: 'x' }), 'edit::index_out_of_range')
-  })
 })
 
 describe('Document editor surface: card mutations', () => {
@@ -1958,173 +1808,5 @@ addr:
     )
     const title = doc.main.payloadItems.find((i) => i.key === 'title')
     expect(title.nestedFills).toBeUndefined()
-  })
-})
-
-// ---------------------------------------------------------------------------
-// reader.resolve, through its `_resolve` ABI: the resolved-value view
-// ---------------------------------------------------------------------------
-//
-// For every declared field: the value the render projection would use and the
-// source rung it came from ("authored" | "default" | "blank"). Rows are an
-// ordered array carrying their own `name`; the card body is a `body` sibling,
-// not a row in `fields`. Value and provenance only: diagnostics stay
-// validate(), guidance stays the schema. See prose/canon/SCHEMAS.md
-// § "Value sources and projections".
-
-describe('reader.resolve, through the _resolve ABI', () => {
-  const QUILL_YAML = `quill:
-  name: field_states_test
-  version: "1.0"
-  backend: typst
-  description: Resolved-field view coverage
-
-main:
-  body:
-    example: "Example body prose."
-  fields:
-    title:
-      type: string
-    status:
-      type: string
-      default: draft
-    notes:
-      type: string
-    count:
-      type: integer
-    author:
-      type: string
-      example: A. Author
-
-card_kinds:
-  note:
-    fields:
-      label:
-        type: string
-`
-
-  const buildQuill = () =>
-    Quill.fromTree(makeQuill({ name: 'field_states_test', quillYaml: QUILL_YAML }))
-
-  // Rows are an ordered array now; look one up by its `name`.
-  const byName = (rows, name) => rows.find((r) => r.name === name)
-
-  it('tags main rows with their authored / default / blank source', () => {
-    const quill = buildQuill()
-    const md = `~~~card-yaml
-$quill: field_states_test
-$kind: main
-title: Hello
-~~~
-`
-    const f = quill._resolve(Document.fromMarkdown(md)).main.fields
-
-    // Declaration order is structural: the array order is the contract.
-    expect(f.map((r) => r.name)).toEqual(['title', 'status', 'notes', 'count', 'author'])
-
-    expect(byName(f, 'title').source).toBe('authored')
-    expect(byName(f, 'title').value).toBe('Hello')
-    expect(byName(f, 'status').source).toBe('default')
-    expect(byName(f, 'status').value).toBe('draft')
-    expect(byName(f, 'notes').source).toBe('blank')
-    expect(byName(f, 'notes').value).toBe('')
-  })
-
-  it('carries the body as a `body` sibling, never a row in fields', () => {
-    const quill = buildQuill()
-    const authored = `~~~card-yaml
-$quill: field_states_test
-$kind: main
-title: T
-~~~
-
-Hello body.
-`
-    const withBody = quill._resolve(Document.fromMarkdown(authored))
-    expect(withBody.main.body).toBeDefined()
-    expect(withBody.main.body.source).toBe('authored')
-    // Not smuggled into the fields array under any `body` / `$body` name.
-    expect(byName(withBody.main.fields, 'body')).toBeUndefined()
-    expect(byName(withBody.main.fields, '$body')).toBeUndefined()
-
-    const blank = `~~~card-yaml
-$quill: field_states_test
-$kind: main
-title: T
-~~~
-`
-    const noBody = quill._resolve(Document.fromMarkdown(blank))
-    expect(noBody.main.body.source).toBe('blank')
-  })
-
-  it('carries value and source only: no diagnostics, no example', () => {
-    const quill = buildQuill()
-    const md = `~~~card-yaml
-$quill: field_states_test
-$kind: main
-title: T
-~~~
-`
-    const f = quill._resolve(Document.fromMarkdown(md)).main.fields
-    // Each row is exactly { name, value, source }, schema guidance (example:)
-    // and diagnostics read from quill.schema / quill.validate, not duplicated.
-    const author = byName(f, 'author')
-    expect(Object.keys(author).sort()).toEqual(['name', 'source', 'value'])
-    expect('example' in author).toBe(false)
-    expect('diagnostics' in author).toBe(false)
-  })
-
-  it('reports kind and index on a card entry', () => {
-    const quill = buildQuill()
-    const md = `~~~card-yaml
-$quill: field_states_test
-$kind: main
-title: T
-~~~
-
-~~~card-yaml
-$kind: note
-label: L
-~~~
-Note body.
-`
-    const states = quill._resolve(Document.fromMarkdown(md))
-    expect(states.cards.length).toBe(1)
-    const card = states.cards[0]
-    expect(card.kind).toBe('note')
-    expect(card.index).toBe(0)
-    expect(byName(card.fields, 'label').source).toBe('authored')
-    expect(byName(card.fields, 'label').value).toBe('L')
-  })
-
-  it('keeps a render-uncoercible value raw (byte-for-byte with the plate)', () => {
-    const quill = buildQuill()
-    const md = `~~~card-yaml
-$quill: field_states_test
-$kind: main
-title: T
-count: "not-a-number"
-~~~
-`
-    // A value the render coercion cannot conform is kept raw and Authored,
-    // exactly as compile_data leaves it: the error surfaces via validate(),
-    // not this view (which carries no diagnostics).
-    const row = byName(quill._resolve(Document.fromMarkdown(md)).main.fields, 'count')
-    expect(row.source).toBe('authored')
-    expect(row.value).toBe('not-a-number')
-    expect('diagnostics' in row).toBe(false)
-  })
-
-  it('result is JSON.stringify-able', () => {
-    const quill = buildQuill()
-    const md = `~~~card-yaml
-$quill: field_states_test
-$kind: main
-title: T
-~~~
-`
-    const states = quill._resolve(Document.fromMarkdown(md))
-    const round = JSON.parse(JSON.stringify(states))
-    expect(byName(round.main.fields, 'title').value).toBe('T')
   })
 })
